@@ -12,22 +12,22 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.0.0",
-    "@angular/common": "^5.0.0",
-    "@angular/compiler": "^5.0.0",
-    "@angular/core": "^5.0.0",
-    "@angular/forms": "^5.0.0",
-    "@angular/http": "^5.0.0",
-    "@angular/platform-browser": "^5.0.0",
-    "@angular/platform-browser-dynamic": "^5.0.0",
-    "@angular/router": "^5.0.0",
+    "@angular/animations": "^5.0.5",
+    "@angular/common": "^5.0.5",
+    "@angular/compiler": "^5.0.5",
+    "@angular/core": "^5.0.5",
+    "@angular/forms": "^5.0.5",
+    "@angular/http": "^5.0.5",
+    "@angular/platform-browser": "^5.0.5",
+    "@angular/platform-browser-dynamic": "^5.0.5",
+    "@angular/router": "^5.0.5",
     "bootstrap": "^4.0.0-beta.2",
     "core-js": "^2.4.1",
     "dat.gui": "^0.6.1",
-    "jquery": "^1.9.1",
+    "jquery": "^1.12.4",
     "ng2-file-upload": "^1.3.0",
-    "popper.js": "^1.12.9",
-    "rxjs": "^5.5.2",
+    "popper.js": "^1.13.0",
+    "rxjs": "^5.5.3",
     "three": "^0.88.0",
     "three-orbit-controls": "^82.1.0",
     "three-ply-loader": "^1.0.0",
@@ -35,9 +35,9 @@
   },
   "devDependencies": {
     "@angular/cli": "1.5.5",
-    "@angular/compiler-cli": "^5.0.0",
+    "@angular/compiler-cli": "^5.0.5",
+    "@angular/language-service": "^5.0.5",
     "@types/dat-gui": "^0.6.3",
-    "@angular/language-service": "^5.0.0",
     "@types/jasmine": "~2.5.53",
     "@types/jasminewd2": "~2.0.2",
     "@types/node": "~6.0.60",
@@ -49,11 +49,11 @@
     "karma-chrome-launcher": "~2.1.1",
     "karma-cli": "~1.0.1",
     "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "~1.1.0",
+    "karma-jasmine": "^1.1.1",
     "karma-jasmine-html-reporter": "^0.2.2",
     "protractor": "~5.1.2",
     "ts-node": "~3.2.0",
     "tslint": "~5.7.0",
-    "typescript": "~2.4.2"
+    "typescript": "^2.6.2"
   }
 }
