@@ -2,25 +2,27 @@ import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 import { HttpModule } from '@angular/http';
 import { HttpClientModule } from '@angular/common/http';
-import { FileSelectDirective } from 'ng2-file-upload';
+import { FileUploadModule } from 'ng2-file-upload';
 
 import { AppComponent } from './app.component';
 import { CameraComponent } from './camera/camera.component';
 import { AppService } from './app.service';
+import { CameraGuiService } from './camera-gui.service';
 
 @NgModule({
   declarations: [
     AppComponent,
-    CameraComponent,
-    FileSelectDirective
+    CameraComponent
   ],
   imports: [
     BrowserModule,
     HttpModule,
-    HttpClientModule
+    HttpClientModule,
+    FileUploadModule
   ],
   providers: [
-    AppService
+    AppService,
+    CameraGuiService
   ],
   bootstrap: [AppComponent]
 })
